Move camera systems to own module
